Latent_Intrinsics-main/requirements.txt
numpy
scipy
pillow
matplotlib
opencv-python
scikit-image
torch>=1.10
torchvision
tqdm